+++ pennylane/__init__.py
@@ -21,7 +21,7 @@ import pkg_resources
 import numpy as _np
 from semantic_version import Spec, Version
 
-from pennylane.queuing import apply_op, QueuingContext
+from pennylane.queuing import apply, QueuingContext
 
 import pennylane.init
 import pennylane.fourier
@@ -61,7 +61,7 @@ from pennylane.utils import inv
 from pennylane.vqe import ExpvalCost, Hamiltonian, VQECost
 
 # QueuingContext and collections needs to be imported after all other pennylane imports
-from .collections import QNodeCollection, apply, dot, map, sum
+from .collections import QNodeCollection, dot, map, sum
 import pennylane.grouping  # pylint:disable=wrong-import-order
 
 # Look for an existing configuration file